Refactoring + Restructuring code, KeyboardWorkaround
- Fixed some minor bugs
- Restructured code so that main doesn't handle EVERYTHING
- Added a workaround for characters that don't get recognized properly (DearPyGui bug)

Characters that aren't properly displayed will be updated to their correct values once the input_text field is unfocused. This currently only works for a couple characters (all Czech characters). I will add more if the need arises or if someone requests them. Hopefully this issue is fixed in the next update of DPG.